Implemented object mapper
