+++ src/SIS.MvcFramework/WebHost.cs
@@ -102,7 +102,7 @@ namespace SIS.MvcFramework
                     Type.GetTypeCode(actionParameter.ParameterType) == TypeCode.String)
                 {
                     var stringValue = GetRequestData(request, actionParameter.Name);
-                    actionParameterObjects.Add(TryParse(stringValue, actionParameter.ParameterType));
+                    actionParameterObjects.Add(ObjectMapper.TryParse(stringValue, actionParameter.ParameterType));
                 }
                 else
                 {
@@ -114,7 +114,7 @@ namespace SIS.MvcFramework
                         var stringValue = GetRequestData(request, propertyInfo.Name);
                         
                         // Convert.ChangeType()
-                        var value = TryParse(stringValue, propertyInfo.PropertyType);
+                        var value = ObjectMapper.TryParse(stringValue, propertyInfo.PropertyType);
 
                         propertyInfo.SetMethod.Invoke(instance, new object[] {value});
                     }
@@ -140,38 +140,6 @@ namespace SIS.MvcFramework
             }
 
             return stringValue;
-        }
-
-        private static object TryParse(string stringValue, Type type)
-        {
-            var typeCode = Type.GetTypeCode(type);
-            object value = null;
-            switch (typeCode)
-            {
-                case TypeCode.Int32:
-                    if (int.TryParse(stringValue, out var intValue)) value = intValue;
-                    break;
-                case TypeCode.Char:
-                    if (char.TryParse(stringValue, out var charValue)) value = charValue;
-                    break;
-                case TypeCode.Int64:
-                    if (long.TryParse(stringValue, out var longValue)) value = longValue;
-                    break;
-                case TypeCode.Double:
-                    if (double.TryParse(stringValue, out var doubleValue)) value = doubleValue;
-                    break;
-                case TypeCode.Decimal:
-                    if (decimal.TryParse(stringValue, out var decimalValue)) value = decimalValue;
-                    break;
-                case TypeCode.DateTime:
-                    if (DateTime.TryParse(stringValue, out var dateTimeValue)) value = dateTimeValue;
-                    break;
-                case TypeCode.String:
-                    value = stringValue;
-                    break;
-            }
-
-            return value;
         }
     }
 }
